+++ src/pysonar_scanner/configuration.py
