+++ src/pysonar_scanner/environment.py
