+++ src/pysonar_scanner/utils/binaries_utils.py
